airflow/requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-apache-spark==4.9.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
numpy==1.26.4
pyspark==3.5.0
